LibWeb: Use doubles for CSS dimension types
Avoid unintentionally converting between float and double multiple times
by just using double everywhere. Also, remove the unused `int` versions
of their constructors.

+++ Userland/Libraries/LibWeb/CSS/Ratio.cpp
@@ -9,7 +9,7 @@
 
 namespace Web::CSS {
 
-Ratio::Ratio(float first, float second)
+Ratio::Ratio(double first, double second)
     : m_first_value(first)
     , m_second_value(second)
 {
